==> Hit Song Prediction & Mood-Based Playlist Classification using spotify/1_eda.ipynb <==
1. Import neccessary library

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

2. Check Dataframe head

df = pd.read_csv('Spotify_Dataset_V3.csv' , sep = ';')
df.head()

3. Check Column

df.columns

4. Check Shape 

df.shape

5. Check Math Metrics

df.describe()

6. Check Data Type

df.info()

7. Check Whether there are nullable values or not

df.isna().sum()

8. Check How Many Unique Song are there ?

# number of total songs
print("Total song:",df['id'].nunique())
