==> Hit Song Prediction & Mood-Based Playlist Classification using spotify/4_datetime.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_csv('delete_duplication.csv' , sep = ';')
df.head()

### Change type Date Column from String to datetime

df.shape

df['Date'] = pd.to_datetime(df['Date'], format='%d/%m/%Y')
print(df['Date'].dtypes)

df['Date']

df.to_csv('change_date_time_column.csv', sep=';', index=False)
