==> Hit Song Prediction & Mood-Based Playlist Classification using spotify/5_loudness_normalization.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import MinMaxScaler


df = pd.read_csv('change_date_time_column.csv' , sep = ';')
df.head()

## Normalize
- Loudness Column

df['Loudness'].describe()


scaler = MinMaxScaler()
df['Loudness_norm'] = scaler.fit_transform(df[['Loudness']])

df['Loudness_norm'].describe()

df.head()

df.to_csv('final_cleaned_data.csv', sep = ';', index = False)
