==> Hit Song Prediction & Mood-Based Playlist Classification using spotify/6_label_data.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_csv("final_cleaned_data.csv", sep = ';' )
df.head()

df['Date'] = pd.to_datetime(df['Date'], errors='coerce')
print(df['Date'].dtypes)

df[df.duplicated(subset=['Title','Date','Artists'], keep=False)]

df['id'].unique()

df_top_50 = df[df['Rank'] <= 50].copy()
df_top_50['first_top_50_date'] = (
    df_top_50.groupby(['id'])['Date']
    .transform('min')
)
song_stay_in_top_50 = df_top_50[df_top_50['Date'] - df_top_50['first_top_50_date'] > pd.Timedelta(days=14)]


song_stay_in_top_50['id'].nunique()
song_stay_in_top_50['id'].nunique() / df['id'].nunique()

 

print(f"number of popular songs: { song_stay_in_top_50['id'].unique().shape[0]} ")
print(f"number of total songs: {df['id'].nunique()}")
print(f"Percentage of popular songs: {(song_stay_in_top_50['id'].nunique() / df['id'].nunique())*100:.2f}%")

### Label column popular in dataset

popular_ids = song_stay_in_top_50['id'].unique()
df['Popular'] = df['id'].isin(popular_ids).astype(int)

df.head()

print(df[df['Popular'] == 1]['Popular'].count())

### Check each song has same popularity

check_pop = df.groupby('id')['Popular'].nunique()
print(check_pop)
inconsistent_songs = check_pop[check_pop > 1]

print(len(inconsistent_songs))

df_song = df.drop_duplicates(subset=['id'], keep='first')

df_song.shape

print("number of popular song",df_song[df_song['Popular']==1]['Popular'].count())
print("number of unpopular song",df_song[df_song['Popular']==0]['Popular'].count())

df_song.to_csv('label_data.csv', sep = ';', index = False)
